Create RecordService class
